backend/api/permissions.py
from rest_framework import permissions

class IsRoleAdmin(permissions.BasePermission):
    def has_permission(self, request, view):
        return bool(
            request.user 
            and request.user.is_authenticated 
            and request.user.role == 'admin'
        )
        
class IsRoleUser(permissions.BasePermission):
    def has_permission(self, request, view):
        return bool(
            request.user
            and request.user.is_authenticated
            and request.user.role == 'user'
        )
        